requirements.txt
streamlit
requests
pytesseract
Pillow